+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.2.1 (Next)
 - Pass marker and keyword expressions as singular strings to command executors.
 - Reorganize module into package.
+- Add testinfra installation verification.
 
 ### 1.2.0
 - Support executing tests by keyword substring expression.